Add new components and dependencies el cual se mejora about
