+++ my-frontend/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "homepage": "https://josee2701.github.io/CV/build",
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.5.1",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
