+++ my-frontend/src/App.css
@@ -15,6 +15,16 @@ body {
   align-items: start;
   flex-direction: column;
 }
+.profile-about-section{
+  inline-size: 100%;
+  max-inline-size: 1200px;
+  margin: 0 auto;
+  padding: 0 20px;
+  display: flex;
+  align-items: start;
+  flex-direction: column;
+
+}
 
 .App-logo {
   inline-size: 150px;
